Support for n-d comprehensions. Try
a = [ rand() | (i=1:3), (j=1:3), (k=1:3) ]

Currently, the code still depends on set() functions generated inside for loops.
set() thus has to be implemented for 1d, 2d, 3d, ... which is less than ideal.
There has to be a way to implement set for arbitrary numbers of inputs.

However, for comprehensions, the next iteration will generate all the index
expressions inlined within the for loop and bypass set() altogether.