1) Correct transformation of comprehensions
Comprehension transformations now do the right thing.
Before this checkin, this would fail for
       x = [ 1 | i=1:2:4 ]
This was because the indexing was blindly substituted, and
the expression was used as is. This of course works for
the cases where the ranges are of the form i=1:n, but fails
for the example above.

Now, the indices in the comprehension evaluation are
correctly transformed. For example
     [ a[I] | I=1:2:4 ]
gets transformed into
     I = 1:2:4
     result = zeros(numel(I))
     for i=1:numel(I)
        result[i] = a[I[i]]
     end

2) Comprehensions no longer require ranges to be in parentheses:
The following is now legitimate
   x = [ 1 | i=1:2:4 ]
whereas earlier, this had to be
   x = [ 1 | (i=1:2:4) ]

3) Comprehension nodes in AST that were tagged as "comp" are
now tagged as "comprehension" for better readability.

4) Scalar and vector reduction operators added.