Fix Makefile so that julia binary doesn't need relinking every time.
This is a bit tricky with the three different ways to link julia, but
my approach is to make differently named binaries for each one and
then just symlink julia to the current one. You still have to clean
and rebuild from scratch if you want a different one, however.