Better way to check that make is catch all dependencies.
This also helps find obsolete files, since non-dependencies are
good canditates for obsolecense. This is a much cleaner and faster
way to check than the check_obsolete.sh script.

Based on that, I've tweaked the Makefile so that now all the
dependencies are captured. Thus it's now safe to just run make any
time any file is changed, since the generated binary should be
remade because of that.