Allow julia to be run from elsewhere: load relative to JUILA_HOME.
If you export the environment variable JUILA_HOME, julia will load
source files relative to this location. Eventually, we'll need a
full load path mechanism, but for now this serves effectively as a
single-directory load path. Export JULIA_HOME and put it in your
PATH and run julia from anywhere.