Experimental technique for avoiding lots of history dupes.
The idea here is that if you scroll back and use a history line
verbatim, instead of creating a new copy of it, you are pinned
back at that point in the history. If you hit up again, you'll get
the same command back. If you hit down, you'll get the command that
was originally executed after it. I think this might be a really
useful way to work.