More numeric support for Chars.
This defaults to always promoting Char to a numeric, then numeric
ops get applied as usual. I decided this since after doing any ops
on a Char, you can't really be sure that it's still a valid char.
This way, all the ops you want can happen, then when you need a
valid Char back, it can get converted and checked for validity.
I.e. we can put all the validity checking in the char() method.