Allow Real indices into Tensors by rounding to the nearest int.
The reasoning behind rounding real indices rather than simply
truncating them to an integers is so that for odd-length vectors
v[end/2] will give the middle element, rather than the element
before the middle. The element before the middle is given in the
case where the vector length is odd since there is no element
precisely in the middle.

This required making a number of the unspecified index parameters
for some of the ref/assign functions more specific, making them of
Indices type, rather than unspecified (cough, Viral, cough).