Bare strings make Latin1String or UTF8String objects by content.
If a bare string encodes a byte sequence that is pure ASCII or
invalid UTF-8, then it creates a Latin1String object. If it is not
pure ASCII and is valid UTF-8 then it creates a UTF8String object.

There is a small chance that someone will enter a series of escaped
bytes using high \x escapes (\x80 and higher), intending it to
become a Latin1String object with bytes corresponding to character
values and accidentally creates a valid UTF-8 byte sequence. In this
case, the programmer may be surprised to get a UTF8String object. To
ensure that a Latin1String object is created, if you're doing
something tricky with high \x escape sequences, you should use the
c"\xE2\x88\x80" form (not yet implemented) since bare "\xE2\x88\x80"
will be interpreted as "\u2200" written out as a sequence of bytes.