Make escape_string match the new string input scheme.
Latin1String objects need to have different escaping behavior from
other strings, since if we output code points between 0x80 and 0xFF
using \u escapes, then the output string would be reinterpreted as
a UTF8String object if input back into the interpreter.

Currently x == eval(quote_string(x)) is not guaranteed since x could
be a Latin1String containing a sequence of bytes which are valid
UTF-8, causing the eval to interpret them as a single unicode char
and making a UTF8String instead of a Latin1String identical to the
original string. However, there is currently no way to input such
strings as literals, which mitigates this problem slightly. In the
future, such strings will have to be output with in c"..." form to
ensure that the x == eval(quote_string(x)) identity holds.