RepString for efficient representation of repeated strings.
I've decided for now to go with * for string concatenation and ^ for
string repetition. Feels quite good as of now. We'll see how it
plays out.

Also rewrote [lr]pad to use * and ^ and thereby get the benefit of
the new efficient string "exponentiation". They are still broken in
that they don't correctly return the right number of characters in 
a variable-width encoding.