Add the rest of the LLVM arithmetic intrinsics and use them.
With this commit, we never use generic definitions for basic
arithmetic types, but call the appropriate LLVM intrinsic directly.
There's even a floating point rem intrinsic, which makes that much
faster. Unfortunately, the fld and mod operations still use the slow
generic code for both int and float types. To ensure that the right
type is generated by this code, I added a convert. I suspect some of
the promotion code needs to be tweaked now.