Use the new gcd behavior to avoid sign multiplication in //.
Even with the new int8 sign values, this would do some unwanted
promotions, e.g. when trying to do uint8(12)//uint8(15), since the
result of Int8*Uint8 is Int16. This is clearly simpler and better.

I also snuck in a test for Rational equality that works even if the
Rational happens to not be in reduced form.