Make tensor.j as abstract as possible.
As of right now, concatenation is the only one where input args are
not Tensors, because the correct specialization does not get picked
otherwise.

I urge everyone to go through this and try to minimize the use of Array.
When satisfactory, we can split off tensor.j and array.j