IPC: very speculative `foo` syntax for "shell" commands.
Downside: would require taking ` back from quoting. On the
other hand, we usually use ` as `begin ... end which is
pretty ugly. Perhaps we should instead have a quote keyword:
quote ... end. The only problem that would leave is symbols.
For symbols, we could introduce another syntax since symbols
and quoting aren't really the same thing: the symbol "foo"
is not actually the same as the expression "foo" at all.
This is currently a bit of a hack, albeit a harmless one.