Simplified (and more correct) sign function for Rational.
This optimization depends on the new presumption that the
Rational representation is always stored in lowest terms
with a non-negative denominator.

Not only is this more efficient, but it also is necessary
now that -1//0, 0//0, and 1//0 are allowed since the old
definition does not agree with what sign(float(r)) would
return.