Implement @assert macro and use it instead of the function.
This looks nicer since you can skip the extra parens before
and after, and allows printing the expression tree when an
assertion fails. It would also potentially allow us to show
the values bound to variables used in the expression, which
would be very helpful in debugging when assertions do fail.