pointer.j: define useful operations with Ptr objects.
This steals a couple of random pointer-related
definitions (pointer equality and WORD_SIZE) and then
adds a bunch of conversions and arithmetic.

I suspect that we should make Ptr <: Int since we're
allowing arithmetic with pointers. This is much like
the situation with Char.

Also: non-promotion-based Int-Char operations. This
is the wrong way to do this since larger Ints might
get truncated to Uint32 and yield wrong ansers.