Combination of Jeff's string literal patch and my additions.
This was tricky to make work and feels pretty brittle, but it
does do the trick. One of the keys here is defining an L"..."
string literal form that just produces its literal string
contents. This is particularly useful in string.j where there
are lots of cases of wanting to generate escape strings for
the printing of escaped representations of strings.