Add r"..." string form that creates a Regex.
Creating Regex objects with regular strings is very awkward
since string escaping clashes with PCRE's quoting. For
example, to escape a . regex metacharacter, you'd have to do
Regex("\\."). For now, r"" just applies the Regex constructor
to the literal string it recieves. Thus this example would
just be r"\.", which far less awkward to write. In the future
I need to make sure that all Julia quoting constructs are
properly handled by this quoting construct. Also of note is
the $ issue: since $ is a regex metacharacter, currently we
don't do any interpolation, but we may want to so something
cleverer here in the future (conditional interpolation).