ShivaIterator: wraps another iterator, created on-demand.
This allows one to create an actual iterator on-demand
only when the iterator is started, as well as calling
a destroy function to clean-up when its done.

What we really need is to augment the iterator protocol
so that you can specify a finalizer method that is
always called when the loop is done, regardless of
how that happens.