Split Functions section into Functions & Methods.
The Functions section can come early -- before types are
introduced. The Methods section requires knowledge of types
and can come later.

I've also commented out sketchy incomplete sections for now.
The text is still there for tweaking, but this makes it more
possible to just give this version of the manual to someone
without confusing the hell out of them.