The sloccount target belongs in top level.
A few other tweaks to the Makefile -- in particular
the julia-{debug,release} targets are real files and
should be treated as such.

Also improved the way the sloccount gets files to
count: now it only gets files that are actually in
the git repository. With all the stuff in ext, we
were tallying in at a whopping $300+ million of
development time. Not quite realistic.