Make external/Makefile significantly more sane.
The following targets now exist for all dependencies:

  compile-$dep install-$dep cleanall-$dep
 
To do all of your platform-specific ones, use:

  compile install cleanall

These all respect timestamps properly. The main issue here is
that it doesn't quite yet interact well with installing in the
toplevel lib directory. I feel like that shouldn't be done from
here since it's bad style for a makefile to mess with anything
above it.