external/Makefile: adding clean-* targets.
The clean-* targets will call clean in the project's
directory and remove the source and target files that
install-* targets create. This is actually much more
involved than just removing the directory altogether,
which can now be accomplished by the distclean-* make
targets. To do clean for all dependencies do cleanall
and to actually remove downloaded files and the dirs
do distclean.