external: systematic naming of tarballs and src dirs.
This patch enforces a consistent naming scheme on all
of the external dependencies:

  tarball: <package>-<version>.<extension>
   source: <package>-<version>/

Among other things, this would make it simple to try
out different versions of external dependencies as
all you would need to do is change the version at
top of the Makefile and run make. To switch back,
just change the version string back.