The dot product should conjugate the second vector.
Matrix multiplication, on the other hand, should not
take a complex conjugate, so shouldn't use the dot
product for its definition -- just use sum(_.*_)
instead. I'm really not sure that dot product makes
any sense for anything that's not a vector so get
rid of the definition for non-vectors. It doesn't
seem to be used anywhere, so this is safe.

Closes #51.