Fix inv(z::Complex); drop norm(z::Complex).
The original definition of complex norm incorrectly
was the square of the absolute value. Jeff fixed it
to be the same as the absolute value, inadvertently
making the definition of inv(z) incorrect. I've done
away with norm for complexes, since abs means exactly
the same thing and norm already has meaning for all
the stuff in linear algebra. However, I've added abs2
which efficiently computes the squared absolute
value of a complex number.

Btw, found this problem by running examples from the
wiki as a manual test suite.