Take these lies and make them true somehow.
Since the creation of Complex64 and Complex128, much
of the manual section on complex numbers had become
a very confusing lie. By introducing a special type
for the "im" global constant, we can make them all
true again, retaining our efficiency but looking sane
once more.

This also serves as an example of how easy it would be
to implement numeric literals (a la Fortress) now.