When pasting indented code into the repl, make it look right.
If the first line is indented, strip that same amount of indent
from the following lines. This behavior is turned off if any of
the input is not "pastable" -- i.e. ^a, ^e, etc. This makes it
so that pasting text that's indented into the repl will repsect
the indentation of the first line, making everything look right.