Part 1: rename {Vector,Matrix} => Abstract{Vector,Matrix}.
This is not a simple search and replace since I'm intentionally
leaving a lot of uses of Vector and Matrix as-is, because I think
that they really should only apply to the dense array types. Thus,
this commit is actually broken: in the next commit, I will drop
the "Dense" prefix from Dense{Vector,Matrix}, completing things.
There are, however, still probably a lot of cases where the dense
forms are more appropriate, but this is a conservative change.