Part 2: rename Dense{Vector,Matrix} => {Vector,Matrix}.
Since this seems really confusing to various people, vector and
matrix now refer to the dense array types, not the abstract tensor
type that they previously referred to. This makes it easier to
"reach" for a Vector or Matrix that are dense and harder to reach
for an AbstractVector or AbstractMatrix, which is good because it
is usually the former that you want, and not the latter.