Better eps(); use hex values for fp traits; hex/oct/bin fp printing.
The previous eps() improvements were part of the story, but not
all of it. See

http://en.wikipedia.org/wiki/Machine_epsilon#How_to_determine_machine_epsilon

for how to do it correctly -- this is what this version uses.

Allowing bin/oct/hex to print floating point values by their
bits makes it much easier to see what's going on with these
operations.

Removing the > and >= methods for floats since these are not
actually necessary: you can just swap the arguments, which is
what the fallback definition does anyway.