Change Regex match API: match returns `nothing` for no match.
This seems cleaner and simpler than returning a useless and
empty RegexMatch object when there is no match. It also makes
the definition of the RegexMatch type simpler since it is no
longer required to be able to indicate that there was not a
match.

Also, change how RegexMatch objects are printed to actually
make it clear that they are special objects, and not just a
weird type of string. Since they don't subtype String, this
is pretty inappropriate. Maybe in the future they could but
for now this seems better.

TODO: update the documentation on regexes to match.