Rational: 0//0 is an error; fixes to rational conversion.
Jeff and I debated this a while ago: his position, divide
by zero when constructing a rational number should be an
error; mine, that rationals should emulate floating-point
behavior. This is a compromise: 1//0 and -1//0 rationals
are ok, but 0//0 is not. Since NaN in most computations is
indicative of an error, this is ok and it simplifies the
rational arithmetic drastically. Having infinite rationals
adds useful functionality without complicating things.

Two fixes to rational conversion: if you try to convert a
floating point value that is bigger than the biggest int
value of the parameter type, it is too big for rationals,
so it should convert as in infinite value. Also, when the
conversion to rational from float happens, it should be
exact, not within eps(x). This is because the comparison
already covers an implicit range dues to floating-point
error. Allowing extra error is wrong unless specifically
requested by the user.