Allow Regex objects to be serialized and deserialized.
The PCRE API allows you to get the size of a regex, and
thereby save and restore it as binary data. Instead of
keeping the regex field of the Regex type as an opaque
pointer, now we keep it as a serializable byte array.
Dealing with the extra field is much harder -- see the
pcreprecompile man page, so we punt and just make regex
object simply not set that field by default. This is a
change in principle, but in practice, I had made a
mistake before and was never using the extra field when
executing a regex anyway, so nothing changes except for
a useless pcre_study call when each regex object is
created.

Addresses the symptom that caused issue #187 to exist.