Efficiency: == for byte strings; cmp for ascii vs utf8.
It is safe to use lexcmp between any pair of ByteStrings;
they don't need to both be ASCIIStrings or UTF8Strings.

Also, for ByteStrings, when checking for equality you can just
compare data lengths first: if they're unequal the strings
cannot possibly be the same, so you can return early. You can
do the same check with strlen between arbitrary strings, but
that would require scanning both of the strings in the general
case, so it's not actually a win.