Fix hashing for large ints and rational numbers.
If an integer cannot be represented exactly as a float, then we
hash its bits instead. Otherwise a lot of contiguous integers
end up in the same hash bin.

Rationals should be hashed the same as ints and floats. This
makes that happen. Use bitmix of the numerator and demoninator
only in the case or rationals that cannot be represented as
ints or as floats.

TODO: there's a bit of an issue remaining here in that hash is
now more correct than isequal in some cases. For example:

  julia> hash(x)
  3931684390250635322

  julia> hash(int(x)+1)
  17732226688739450744

  julia> isequal(x,int(x)+1)
  true

In other words, hash can differentiate these but isequal can't!
This is because isequal unconditionally promotes when doing a
comparison. We could add specific isequal methods but that has
an O(n*m) method explosion. Another approach is to make the
isequal method compare hashes first and then check for some
other kind of equality. This doesn't have a explosion problem
since we only need to make sure hashing is good for each type.