Back to honest printing, except for in arrays.
What a relief:

  julia> nextfloat(1.0)
  1.00000000000000022

  julia> prevfloat(1.0)
  0.99999999999999989

This really bothered me because these used to both look the
same as 1.0 and they're very much not.

In arrays showcompact is used instead, which does what is
advertised by the name. I'm playing around a bit with how
to show complex numbers in arrays too. Trying out the (r,i)
notation to see how it feels. Still need to figure out how
to handle aligning complex numbers gracefully, but I have
an idea for that.