Floating-point intrinsics for isequal* and isless.
*Note: this intrinsic doesn't handle making all NaNs equal.
For example, isequal(NaN,-NaN) fails with this. I've looked
into it a bit and cannot figure out a fast, clever way to
do the comparison. Frustratingly, the fcmp operations that
LLVM provides don't do what's needed. There are 16 of them,
none is what is needed.

The isless comparison, however, is a fast way to order the
floating-point values so that they sort nicely, with NaNs
last as they are in Matlab.