Implement conj! and make ctranspose = transpose for non-numerics.
When you have a non-numeric vector or matrix, you want to be able
to write X' and have it not barf when trying to conjugate, say,
strings. One way around this would be to define a default conj
that is just the identity and only define conj as non-identity on
non-real numbers. However, that seems a bit kludgy. Really, conj
should just never be called if the element type is non-numeric.