Make ``run(`cmd`)`` expect success; use ``success(`cmd`)`` to test for it.
Most of the time we mean the former and should fail if the called
command fails -- but we're lazy and don't test for it. This way we
can be lazy and still have safe programs. If you really don't care
about failure, just use ``success(`cmd`)`` and ignore the result.