Rename: long => int, ulong => uint.
This is probably the semantically diciest change because it kind
of clashes with the C meaning of int and uint, which is 32-bit on
all (relevant) platforms. However, these are conversion functions
in Julia anyway, so it's probably not a big issue. Just might be
confusing. In Julia if you want to produce a 32-bit int, use int32
and if you want to produce a system-sized integer use int.