Rename clock() to time().
This function doesn't really work like the Matlab function
called clock() but does work like the C function called
time() so it seems least confusing for every potential user
for this to be called time() instead of clock(). We could
have a clock() function that works like the Matlab one.