Deterministic average-of-three quicksort pivot selection.
This doesn't lose performance relative to using the midpoint in the
sorting benchmark and avoids the common worst-case scenarios when
the array is already sorted or sorted in reverse. This is probably
pretty competetive with timsort, to be honest, since timsort is
designed to work on generic elements rather than purely numeric
elements.