Faster subprocess spawning.
The issue here was that we were trashing the memory space of the
forked child process before doing an exec. This causes copy-on-
write in the kernel for the child's memory pages and puts the child
process off the "happy path" that just goes straight from fork to
exec without descheduling the child first. We're still not as fast
as a shell fork and exec, so I'm not sure what's up with that, but
at least we're better than we were.