wip: using the "padding" bit method for eq_f64_u64 code gen too.
I'm not sure if this approach is better since it's much more
complicated, but it does potentially allow LLVM to do constant
folding when the integer is a computed value and determine when it
compiles the code that the integer part doesn't need to be done at
run time. Mainly checking it in so that we can have it preserved.

This method doesn't work as nicely for signed integers because of
the sign bit. Basically, you need to take the absolute value, but
I don't want to introduce branches.