Implement all mixed 64-bit integer/float comparisons.
This includes a different way of checking equality that is
just as fast, but allows LLVM to take care of constant
folding for us since the equality of an integer to itself
after mapping to floating-point and back can be evaluated
statically in the cases where the integer is constant. This
applies even in cases where it is a constant-foldable
expression rather than an actual constant.

Better still, these < and <= operations do not, by trivial
benchmarking at least, appear to be any slower than what we
were doing before, hard as that is to believe. They seem to
play quite nicely with the CPU pipeline.