Better length calculation for non-integer ranges.
This is admittedly a pretty esoteric case, but this affects, e.g.
-2.0^52:2.0^52 since adding 1 to 2^53 has no effect. This way the
one actually gets added first. It needs to be ifloor rather than
itrunc because otherwise 0:-0.5 ends up having a lenght of 1.

To make this work for rationals requires defining itrunc, ifloor,
and iceil for them, which easily allows defining trunc, floor and
ceil as well.