Don't promote xor ($), a.k.a "Int arithmetic part 2".
Since we now promote to 64 bits regardless of the sizes of the
arguments, it's a bad idea to promote xor operations since the
result is now quite different. This exposed a couple of cases
in intfuncs.j where we are now xoring values that previously
would have been the same size but now one was promoted to Int.

I considered exposing modular addition for each integer type,
but decided to just use intrinsics where necessary instead.