wip: use Grisu algorithm for efficient optimal float printing.
This algorithm is great, as is the implementation, but the
"library" is insanely hard to include. This is not friendly to
use in other projects at all. Nevertheless, it works.

Basic usage:

  grisu(x)
  grisu(x,n)
  grisu(x,-n)

Returns `(sign, digits, point)` where `sign` is boolean and
indicates the sign of the number, `digits` is an ASCII string
of the actual decimal digits, and `point` is an integer offset
indicating where relative to the start of `digits` the decimal
point should go (it can be zero, positive or negative). The
forms with `n` give fixed precision (number of significant
digits) when `n` is positive, and fixed trailing digits (after
the decimal point) when `n` is negative.

Examples:

  julia> load("j/grisu.j")

  julia> grisu(1)
  (true,"1",1)

  julia> grisu(1.25)
  (true,"125",1)

  julia> grisu(0.1)
  (true,"1",0)

  julia> grisu(1/7)
  (true,"14285714285714285",0)

  julia> grisu(float32(1/7))
  (true,"14285715",0)

  julia> grisu(1/7, 8)     ## 8 significant digits
  (true,"14285714",0)

  julia> grisu(100/7, 8)   ## 8 significant digits
  (true,"14285714",2)

  julia> grisu(100/7, -8)  ## 8 trailing digits
  (true,"1428571429",2)