Fix symptoms of possible grisu/ccall bug.
In the ccall on line 11, I'm using the single-element sign::Array{Int32}
buffer to get a boolean flag set by the call to libgrisu. However, this
was getting these bit-patterns:

  julia> grisu(1.2)
  (131072,"12",1)

  julia> bin(131072,32)
  "00000000000000100000000000000000"

  julia> grisu(-1.2)
  (131073,"12",1)

  julia> bin(131073,32)
  "00000000000000100000000000000001"

So there's a bit set in the high word of the Int32 in sign as well. The
commit avoids this problem, but it may be indicative of a bigger issue.

Also: in print_shortest, if k > 2, we always want the exponential
notation, checking whether k = 2 or k = 3 is overkill.