A more traditional interface to printf.
You can use printf, sprintf like this:

  julia> printf("| %-8s | %3d |\n", "foo", 3)
  | foo      |   3 |

  julia> sprintf("| %-8s | %3d |\n", "foo", 3)
  "| foo      |   3 |\n"

For extra efficiency, however, you can use a prefixed format
string, starting with an "f":

  julia> printf(f"| %-8s | %3d |\n", "foo", 3)
  | foo      |   3 |

  julia> sprintf(f"| %-8s | %3d |\n", "foo", 3)
  "| foo      |   3 |\n"

The f"..." format string creates a function that will print
its arguments formatted when applied (assuming they are of the
correct number and type). The printf function simply applies
this function to the rest of its arguments. The sprintf func-
tion does the same but captures the result to a string. There
are also variants of printf that take an IOStream as the first
argument and print to it.