Add libgrisu to the install target.
Update the Makefile's install target to not use rm, cp, etc. and
instead use the install command.

The default install path is $(DESTDIR)/usr/share/julia. In case
DESTDIR is not specified during make, the installation happens in
/usr/share/julia.

I would have liked this to be /usr/local/julia instead, but the debian
package creation scripts do not like this, and it calls the install
target in the Makefile to create the package. I think /usr/share/julia
is acceptable as a default for the time being.

I believe this sufficiently addresses issue #6 even though the solution
is not completely satisfactory.