Fast unsigned integer octal and hex decoding.
These write only the significant digits of an unsigned integer into
the global buffer _jl_digits and set _jl_length[1] to the number of
significant digits there and _jl_point[1] to the offset of the
"decimal" point relative to the first digits. This is similar to
the way that the grisu library does, providing a consistent way to
express decoded numeric values.