printf refactor: fast consistent integer & float decoding.
This brings the speed of the printf("%d %d") benchmark down to the
108 ms range, compared to C's 22 ms. The big slow-down here seems
to be all the write calls, so the next move is to reduce how many
of those there are by using a buffer to build the printf output.

After that, we have to reduce the overhead of the write call also,
possibly by getting rid of the try-catch construct.

Consistency refers to the fact that after decoding, there is no
difference between integers and floats from the perspective of the
formatting machinery.