Replace the ever-confusing `empty(String)` with `String[]`.
It's confusing because `empty(String)` doesn't construct the
empty string. It was for constructing an empty array of type
Array{String,1}. This is so much clearer. It's also more
powerful since you can now write things like this:

   String["foo","bar","baz"]

There was no way to construct this array in one line before.