Fixing a method sort order bug that caused this ordering:
==(Rational{T<:Integer},Number) at rational.j:105
==(T<:Number,T<:Number) at promotion.j:75
==(Rational{T<:Integer},Integer) at rational.j:98

which in turn caused issue #354 together with the recent changes to
unsigned integer arithmetic.
This is tricky because while (Rational,Integer)<:(Rational,Number),
(Rational,Number) has overlap with (T,T) but (Rational,Integer) doesn't.

The new sort order is better, and exposed 2 new method ambiguities in
rational which are also fixed here.