Process improvements.
Catch "deep" errors after fork indicated by 0xff status. This is a
bit sketchy since a process could also return 0xff as its status,
but this allows us to use ignorestatus to ignore non-zero returns
but still catch when something worse goes wrong, like the command
to be executed not existing at all.

Also, correctly distinguish between "already running" and "already
run" when complaining about trying to start processes that are not
in the not-yet-run state.