Add -fno-strict-aliasing to CFLAGS to fix a unittest fail in FreeBSD amd64
Since there are plenty of pointer casts all over the code strict
aliasing is currently not observerd.  This can lead to hard to find
and debug errors.

Once all pointer casts and other pointer aliasing is for sure removed
from the code this flag can be removed again.  Or if the small
performance loss in gcc compiled code is not a problem it can also be
left in.

See thsi very good article for explanation of the problem:
http://cellperformance.beyond3d.com/articles/2006/06/understanding-strict-aliasing.html