More futzing with build version numbering.
I realized that leading zeros would get stripped from the time
parts of the timestamp. Since unix time is just a number, it's ok.
It's also a very compact ordered representation of the date. To
avoid git hashes being interpreted as numbers when they happen to
only have digits, prefix the hash part with "r". To indicate a
dirty build, use a separate build component, "dirty". This is good
because we want this to be easily noticable.