Rename: examples/priority_queue.jl => examples/heap.jl.
This is what the data structure defined in the file is called, so
it's a better name for the file itself. Also, when I read "priority
queue", it takes me seconds to comprehend what we're talking about
whereas I have immediate recognition of what a head data structure
is (there's some confusion with "the heap", but that's minimal).

The number output by typemin(Int64) is now an error if put back in
as a literal:

    julia> typemin(Int64)
    -9223372036854775808

    julia> -9223372036854775808
    syntax error: invalid numeric constant 9223372036854775808

However, this is arguably an improvement since previously instead
of giving an error, this just gave you a value of the wrong type:

    julia> -9223372036854775809
    0x7fffffffffffffff

It also makes the manual a paragraph shorter, which is usually a
good sign.