Filter and Reverse iterators; functions fall back on these.
As per this discussion:

https://groups.google.com/d/topic/julia-dev/uOZoMtabElE/discussion

the uppercase constructors always create an iterator object,
wrapping another iterable to alter how it is iterated. The
lowercase version falls back on this as a default, but does
something eager whenever possible, i.e. actually reverse a
vector or a string or what have you.