Make optional strchr arg a starting index (1-based), not an offset.
Sorry for the noise, folks. I decided that having this as a 0-based
offset was a misdesign. It got awkward to program with pretty fast.
Now the third argument for strchr and match and such is always an
index into the string to be searched, at which the search begins.