wip: package management — a version using a single master branch
This makes each repo commits parents include the submodule commits,
allowing a single branch to automatically root everything. The bad
is that it makes the history look super weird. The good is that if
you rewrite history by collapsing commits, the rooting of anything
that is elided is automatically removed.