invperm: avoid expensive isperm check.
There are two (non-exclusive) ways in which the input can fail to
be a permutation:

1. it contains a number outside of the range 1 through `length(a)`
2. it contains some number in the range `1:length(a)` more than once

In the former case, you'll get a indexing error; in the latter case,
you'll have some element in `b` which never gets assigned.

Sorting the permutation is an O(n log n) check. This way of checking
lazily that the input is a permutation is O(n), so somewhat better.
The bigger win is really that it doesn't require creating an extra
copy of the array just to sort it.