Distributions: some initial type definitions.
The list of distributions is a (small) subset of what SciPy has:

    http://docs.scipy.org/doc/scipy/reference/stats.html

I've taken the parameters of the distrubutions from wikipedia and
thus far only included distributions that have wikipedia pages.
This differs from the SciPy approach, which tries to use location
and scale parameters uniformly. That approach can be quite surpr-
ising to statisticians who are accustomed to more traditional
parameters for these distrubtions.