Tried a somewhat nicer implementation of the LLVM dump.
The builtin is visible to julia as llvm_dump_lambda and needs a
LambdaStaticData object as input. The front-end llvm_dump gets
this object for a given method and calls the builtin. The llvm
code is returned as a string, which can then be printed, saved to
a file, etc.