doc: fix a number of fragment specifiers.
In general, the anchors and fragments of the restructured text
"translation" of the documentation seems to be a total mess. Some
of the anchors are `#id1`, which is not desirable. Some anchors
have a `man-` prefix, while others don't. I have no idea why the
tool generates these anchors differently.

There are also a lot of broken references in the text. These seem
to have been poorly translated by some automated tool.