Fix for printing typemin(Int32); extensive int printing tests.
The core fix here is making the unsigned() function preserve
the size of its argument, which is safe since it is as value-
preserving as the function can be.

This commit also renames int2str as base and moves the base
argument to the first position. We should probably make a
corresponding change to parse_int, but I'll leave that for
another commit.