base: allow using arbitrary ASCII symbols for integer output.
If you want non-ASCII symbols, you'll have to generate ASCII
and then translate and transcode to UTF-8. That's beyond the
scope of printing integers, imo. This is already more generic
than is probably necessary, but it is kind of fun that you can
do things like generate strong ASCII passwords like this:

  julia> base(62,randi(Uint64))
  "cZkH0mnXADB"

  julia> base(62,randi(Uint64))
  "1WgQjGFTDOU"

  julia> base(62,randi(Uint64))
  "aSPs2nMsJAk"