version: allow the trailing pre-release identifier to be empty.
Also, sort the empty string before all other identifiers, thereby
making the version v"1.2.3-" come before all v"1.2.3" versions.
This will be handy when we start specifying version ranges in the
package manager and want to be able to stop a version range before
any variant of some version, including any and all pre-releases.

Something similar could be done with build versions, but it would
be more complicated because you'd want the empty identifier to be
*after* all other identifiers, and less useful because typically
you have a first-known good version that serves as a closed end-
point for your interval of acceptable versions.