Utility functions for efficient FFTs
FFTs are most efficient if the length is a product of powers of small primes. nextpow2 allows you to compute the next power-of-two, and is a good choice for 1 dimensional FFTs. However, for 3-dimensional FFTs, using nextpow2 could result in an array that contains almost 8-fold more elements, and in this case the efficiency of the FFT algorithm does not usually compensate for the extra elements.

These utilities allow you to define something like "nextpowb", where b is a list of possible bases.  Example:

    # compute all integers that are 2^l*3^m*5^n, up to first that is >= 1025
    p = powers([2,3,5],1025)
    s = nextsize(220, p)