Julia types for CHOLMOD structs
Added Julia types CholmodCommon, CholmodSparse, CholmodFactor,
CholmodDense and CholmodDenseOut to extras/linalg_suitesparse.jl.
CholmodSparse and CholmodDense allocate the cholmod_* struct and
populate it with pointers to the contents of a Julia object.  They are
read-only in the CHOLMOD code.  CholmodCommon is a C struct populated
by Julia.  CholmodFactor contains a pointer generated by the C code
and the fields accessed through the pointer are allocated by the C
code.  It is opaque in Julia.  CholmodDenseOut (and the to-be-created
CholmodSparseOut type) are types for output from compiled CHOLMOD
code.  They contents pointed to in these structs are allocated and
controlled by the C code.  Thus the finalizer for a CholmodDenseOut
instance calls the C function :cholmod_l_free_dense but the finalizer for
a CholmodDense instance only calls _c_free on the pointer to the
struct.

When the dense or sparse result of a CHOLMOD function is to be used in
Julia a copyout method is called on it to generate the Julia object.
There are two stages: first the type and dimensions of the Julia
object are determined then the object is created and passed to a C
function that copies the contents of the CHOLMOD storage into it.
Right now this last part isn't working as I am not getting the name of
a type in the correct form in the copyout method for
CholmodDenseOut{T}.  The type is determined dynamically after the call
to jl_cholmod_dense_out_size but the subsequent call to
jl_cholmod_dense_copy_out is malformed.

In any case I wanted to get this committed for Viral's comments.  It
still passes its tests, but that is because there are no tests :-)