Extend Cholmod* types and methods
Defined new Cholmod* types and added methods for solve, size, nnz,
etc.  The C function jl_cholmod_dense_out_size is no longer needed as
the size is now stored in the Julia object.

The CholmodPtr type, which should be amalgamated with the UmfpackPtr
type, carries along the value and index types.