Major libuv interface refactor
This commit refactors both the C side and the Julia side of the libuv interface, moving data from C structures to Julia structures, where possible.

This is achieved by introducing a 1:1 correspondance between Julia and C objects (as discussed on IRC) and replacing custom callbacks passed to C function, by only one Julia function (or multiple, utilizing multiple dispatch), which can then in turn call function stored in the Julia type (to provide custom callbacks).

This refactor also made it easy to integrate libuv with the gc, so that we are not leaking handles anymore and do not need to utilize hacks like the `make_callback` function.