Allow 32-bit and 64-bit CHOLMOD objects on 64-bit systems
CHOLMOD has a peculiar templating system which I misunderstood in
earlier versions of this code.  Changing from the cholmod_l_*
functions to the cholmod_* functions and using 32-bit integers where
needed allows for CHOLMOD operations on both 32-bit and 64-bit integer
sparse matrices on a 64-bit system.

I also changed the conversion routine that populates a SparseMatrixCSC
object from the contents of a CholmodSparseOut object so that it
handles unsorted Cholmod objects and objects with itype of
CHOLMOD_INTLONG more cleanly.