Change convert methods to external constructors
In many cases there is little to distinguish an outer constructor that
takes a single argument of a given type from a convert method.  For
the types in this file a user must specify the target type completely
for a convert method whereas only the type template name must be
specified for an outer constructor.