det based on LU decompostion, not QR
The previous definition of det using the qr() function would sometimes
get the sign wrong.  It's easiest to get the correct sign when using
an LU decomposition.  Also added some tests.  It appears that there
may be problems with eig applied to a Hermitian matrix so that test is
currently suppressed.