show errors: indicate when the problem is with show and not eval.
This is a common source of confusion when working in the repl with
new types that can't show themselves successfully. The repl show
functionality is pretty darned convoluted now -- I couldn't work
out how to make the error show up with a backtrace. This stuff
could use some refactoring to make it more understandable.