fixing regression in reshape() that was causing instability.
the interaction between reshape() and growing arrays is tricky. an array
that never grows should never have its data move. but, if its data is shared
by reshape() and you try to grow the array later, we are in trouble because
then it is too late to change the representation. solved for now by copying
small arrays on reshape().