Add performance test for nested looping
This is extremely curious: in a nest of 3 loops, I get a huge
performance gain (10x) by moving the inner loop into a separate
function. If I had to speculate wildly, I suspect this has something
to do with the compiler worrying that the outer loop variables might
be changed by the inner loop? And that putting it inside a function
removes that concern?